fix(registry): isolate cliproxy plugin entrypoint
